feat(v1-10): side panels visual coherence — PanelHeader unification & decomposition
## 变更概述

统一 5 个侧边面板的视觉语言，提取 PanelHeader 共享组件，将超大组件拆分至 ≤300 行。

## 关键实现

### Part A: PanelHeader 共享组件
- 新增 `components/patterns/PanelHeader` — h-10, flex, px-3, border-b separator
- Props: title, subtitle?, actions?
- 5 个面板的头部统一使用此组件（AC-1）

### Part B: 面板分解（AC-24，全部 ≤300 行）
| 面板 | 重构前 | 重构后 |
|------|--------|--------|
| KnowledgeGraphPanel | 1305 | 147 |
| OutlinePanel | 1093 | 298 |
| CharacterDetailDialog | 1090 | 264 |
| MemoryPanel | 910 | 145 |
| VersionHistoryContainer | 423 | 272 |
| CharacterPanel | 379 | 225 |
| VersionHistoryPanel | 183 | 183 |

提取新文件：kgTypes/Utils/Actions/useKgPanelState/KgEntityCard/KgRelationCard/KgListView/ViewModeToggle、OutlineNodeItem/Tree/Types/useOutlineDrag、CharacterBasicInfo/Relationships/Appearances/character-detail-shared、MemoryCard/Editor/List/useMemoryState/memoryPanelTypes、CharacterPanelSections、versionHistoryHelpers（convertToTimeGroups 提取）

### Part C: eslint-disable 清理（AC-18）
- 主线 baseline: 61 条 → 29 条（减少 53%）
- 剩余 29 条均为合法保留：label/radio(无 Label primitive)、textarea(无 forwardRef)、复杂条件样式 Popover 按钮
- Guard 测试阈值设为 ≤30（evidence-based）

### Part D: 守卫测试
- 新增 `panel-visual-coherence.guard.test.ts`（13 测试）
  - AC-1: 5 个面板文件均使用 PanelHeader ✅
  - AC-24: 7 个主文件行数 ≤300 ✅
  - AC-18: eslint-disable ≤30 ✅
- 修复 tooltip-title-guard：PanelHeader 加入 HEADING_TITLE_COMPONENTS 白名单

## 验证
- 测试：2316/2316 通过（100%）
- typecheck：零错误
- Storybook：构建成功

Closes #1181
