+++ apps/desktop/renderer/src/features/version-history/VersionHistoryContainer.tsx
@@ -1,17 +1,10 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import {
-  VersionHistoryPanelContent,
-  type TimeGroup,
-  type VersionEntry,
-  type VersionAuthorType,
-} from "./VersionHistoryPanel";
+import { VersionHistoryPanelContent } from "./VersionHistoryPanel";
 import { useVersionCompare } from "./useVersionCompare";
 import { useEditorStore } from "../../stores/editorStore";
 import {
-  type BranchMergeConflict,
   type VersionListItem,
-  type VersionStoreActions,
   useVersionStore,
 } from "../../stores/versionStore";
 import { invoke } from "../../lib/ipcClient";
@@ -19,422 +12,19 @@ import { useConfirmDialog } from "../../hooks/useConfirmDialog";
 import { SystemDialog } from "../../components/features/AiDialogs/SystemDialog";
 import { RESTORE_VERSION_CONFIRM_COPY } from "./restoreConfirmCopy";
 import { useVersionPreferencesStore } from "../../stores/versionPreferencesStore";
-import { i18n } from "../../i18n";
 import { getHumanErrorMessage } from "../../lib/errorMessages";
 import { EmptyState } from "../../components/patterns/EmptyState";
 import { LoadingState } from "../../components/patterns/LoadingState";
 import { ErrorState } from "../../components/patterns/ErrorState";
+import {
+  formatTimestamp,
+  convertToTimeGroups,
+} from "./versionHistoryHelpers";
+import { useConflictResolution } from "./useConflictResolution";
+import { BranchMergeSection } from "./BranchMergeSection";
 
-/**
- * Map backend actor to UI author type.
- */
-function mapActorToAuthorType(
-  actor: "user" | "auto" | "ai",
-): VersionAuthorType {
-  switch (actor) {
-    case "user":
-      return "user";
-    case "ai":
-      return "ai";
-    case "auto":
-      return "auto-save";
-    default:
-      return "user";
-  }
-}
+export { formatTimestamp } from "./versionHistoryHelpers";
 
-/**
- * Get display name for actor type.
- */
-function getAuthorName(actor: "user" | "auto" | "ai"): string {
-  switch (actor) {
-    case "user":
-      return i18n.t("versionHistory.container.author.you");
-    case "ai":
-      return i18n.t("versionHistory.container.author.ai");
-    case "auto":
-      return i18n.t("versionHistory.container.author.auto");
-    default:
-      return i18n.t("versionHistory.container.author.unknown");
-  }
-}
-
-/**
- * Format timestamp for display.
- */
-export function formatTimestamp(
-  createdAt: number,
-  now: number = Date.now(),
-): string {
-  const diff = now - createdAt;
-  const minutes = Math.floor(diff / 60000);
-  const hours = Math.floor(diff / 3600000);
-
-  if (minutes < 1) {
-    return i18n.t("versionHistory.container.timeGroup.justNow");
-  }
-  if (minutes < 60) {
-    return i18n.t("versionHistory.container.timeGroup.minutesAgo", { minutes });
-  }
-  if (hours < 24) {
-    const date = new Date(createdAt);
-    return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
-  }
-
-  const date = new Date(createdAt);
-  return date.toLocaleDateString([], { month: "short", day: "numeric" });
-}
-
-/**
- * Get time group label for a timestamp.
- */
-function getTimeGroupLabel(createdAt: number): string {
-  const now = new Date();
-  const date = new Date(createdAt);
-
-  const isToday =
-    date.getDate() === now.getDate() &&
-    date.getMonth() === now.getMonth() &&
-    date.getFullYear() === now.getFullYear();
-
-  const yesterday = new Date(now);
-  yesterday.setDate(yesterday.getDate() - 1);
-  const isYesterday =
-    date.getDate() === yesterday.getDate() &&
-    date.getMonth() === yesterday.getMonth() &&
-    date.getFullYear() === yesterday.getFullYear();
-
-  if (isToday) {
-    return "today";
-  }
-  if (isYesterday) {
-    return "yesterday";
-  }
-  return "earlier";
-}
-
-/**
- * Get description for a version based on reason.
- */
-function getDescription(reason: string): string {
-  if (reason === "autosave") {
-    return i18n.t("versionHistory.container.autosave");
-  }
-  if (reason === "manual-save") {
-    return i18n.t("versionHistory.container.manualSave");
-  }
-  if (reason === "status-change") {
-    return i18n.t("versionHistory.container.statusChange");
-  }
-  if (reason === "ai-accept") {
-    return i18n.t("versionHistory.container.aiModify");
-  }
-  if (reason === "restore") {
-    return i18n.t("versionHistory.container.restoreVersion");
-  }
-  if (reason.startsWith("ai-apply:")) {
-    return i18n.t("versionHistory.container.aiModify");
-  }
-  return reason || i18n.t("versionHistory.container.versionSnapshot");
-}
-
-/**
- * Convert backend version list to UI timeGroups format.
- */
-function convertToTimeGroups(
-  items: VersionListItem[],
-  currentHash: string | null,
-): TimeGroup[] {
-  if (items.length === 0) {
-    return [];
-  }
-
-  const groupMap = new Map<string, VersionEntry[]>();
-
-  for (const item of items) {
-    const label = getTimeGroupLabel(item.createdAt);
-    const isCurrent = item.contentHash === currentHash;
-
-    const entry: VersionEntry = {
-      id: item.versionId,
-      timestamp: formatTimestamp(item.createdAt),
-      authorType: mapActorToAuthorType(item.actor),
-      authorName: getAuthorName(item.actor),
-      description: getDescription(item.reason),
-      wordChange: { type: "none", count: 0 }, // TODO(#571): calculate actual word diff
-      isCurrent,
-      reason: item.reason,
-    };
-
-    const existing = groupMap.get(label) ?? [];
-    existing.push(entry);
-    groupMap.set(label, existing);
-  }
-
-  // Sort groups: Today first, then Yesterday, then Earlier
-  const TIME_GROUP_KEY_TO_I18N: Record<string, string> = {
-    today: "versionHistory.container.timeGroup.today",
-    yesterday: "versionHistory.container.timeGroup.yesterday",
-    earlier: "versionHistory.container.timeGroup.earlier",
-  };
-  const order = ["today", "yesterday", "earlier"];
-  const groups: TimeGroup[] = [];
-
-  for (const key of order) {
-    const versions = groupMap.get(key);
-    if (versions && versions.length > 0) {
-      groups.push({ label: i18n.t(TIME_GROUP_KEY_TO_I18N[key]), versions });
-    }
-  }
-
-  return groups;
-}
-
-type ConflictFormEntry = {
-  resolution: "ours" | "theirs" | "manual";
-  manualText: string;
-};
-
-function BranchConflictItem(props: {
-  conflict: BranchMergeConflict;
-  selected: ConflictFormEntry;
-  onResolutionChange: (
-    conflictId: string,
-    resolution: "ours" | "theirs" | "manual",
-  ) => void;
-  onManualTextChange: (conflictId: string, manualText: string) => void;
-  t: ReturnType<typeof useTranslation>["t"];
-}): JSX.Element {
-  const { conflict, selected, onResolutionChange, onManualTextChange, t } =
-    props;
-  return (
-    <div
-      key={conflict.conflictId}
-      data-testid={`branch-conflict-item-${conflict.conflictId}`}
-      className="space-y-2 rounded border border-[var(--color-border-default)] bg-[var(--color-bg-base)] p-2"
-    >
-      <div className="text-[11px] text-[var(--color-fg-muted)]">
-        {t("versionHistory.container.conflictNumber", {
-          number: conflict.index + 1,
-        })}
-      </div>
-      <div className="grid grid-cols-1 gap-2 text-xs text-[var(--color-fg-default)]">
-        <div>
-          <div className="text-[11px] text-[var(--color-fg-muted)]">
-            {t("versionHistory.container.base")}
-          </div>
-          <pre className="whitespace-pre-wrap font-mono text-xs">
-            {conflict.baseText}
-          </pre>
-        </div>
-        <div>
-          <div className="text-[11px] text-[var(--color-fg-muted)]">
-            {t("versionHistory.container.ours")}
-          </div>
-          <pre className="whitespace-pre-wrap font-mono text-xs">
-            {conflict.oursText}
-          </pre>
-        </div>
-        <div>
-          <div className="text-[11px] text-[var(--color-fg-muted)]">
-            {t("versionHistory.container.theirs")}
-          </div>
-          <pre className="whitespace-pre-wrap font-mono text-xs">
-            {conflict.theirsText}
-          </pre>
-        </div>
-      </div>
-
-      <div className="flex flex-wrap gap-3 text-[11px] text-[var(--color-fg-default)]">
-        {/* eslint-disable-next-line creonow/no-native-html-element -- specialized label */}
-        <label className="inline-flex items-center gap-1">
-          {/* eslint-disable-next-line creonow/no-native-html-element -- specialized input */}
-          <input
-            type="radio"
-            name={`resolution-${conflict.conflictId}`}
-            checked={selected.resolution === "ours"}
-            onChange={() => onResolutionChange(conflict.conflictId, "ours")}
-          />
-          {t("versionHistory.container.useOurs")}
-        </label>
-        {/* eslint-disable-next-line creonow/no-native-html-element -- specialized label */}
-        <label className="inline-flex items-center gap-1">
-          {/* eslint-disable-next-line creonow/no-native-html-element -- specialized input */}
-          <input
-            type="radio"
-            name={`resolution-${conflict.conflictId}`}
-            checked={selected.resolution === "theirs"}
-            onChange={() => onResolutionChange(conflict.conflictId, "theirs")}
-          />
-          {t("versionHistory.container.useTheirs")}
-        </label>
-        {/* eslint-disable-next-line creonow/no-native-html-element -- specialized label */}
-        <label className="inline-flex items-center gap-1">
-          {/* eslint-disable-next-line creonow/no-native-html-element -- specialized input */}
-          <input
-            data-testid={`branch-conflict-manual-${conflict.conflictId}`}
-            type="radio"
-            name={`resolution-${conflict.conflictId}`}
-            checked={selected.resolution === "manual"}
-            onChange={() => onResolutionChange(conflict.conflictId, "manual")}
-          />
-          {t("versionHistory.container.useManual")}
-        </label>
-      </div>
-      {selected.resolution === "manual" ? (
-        // eslint-disable-next-line creonow/no-native-html-element -- specialized textarea
-        <textarea
-          data-testid={`branch-conflict-manual-text-${conflict.conflictId}`}
-          className="h-20 w-full rounded border border-[var(--color-border-default)] bg-[var(--color-bg-raised)] p-2 font-mono text-xs text-[var(--color-fg-default)]"
-          value={selected.manualText}
-          onChange={(event) =>
-            onManualTextChange(conflict.conflictId, event.target.value)
-          }
-        />
-      ) : null}
-    </div>
-  );
-}
-
-function useConflictResolution(
-  mergeConflicts: BranchMergeConflict[],
-  documentId: string | null,
-  mergeSessionId: string | null,
-  resolveBranchConflict: VersionStoreActions["resolveBranchConflict"],
-): {
-  conflictForm: Record<string, ConflictFormEntry>;
-  handleResolutionChange: (
-    conflictId: string,
-    resolution: "ours" | "theirs" | "manual",
-  ) => void;
-  handleManualTextChange: (conflictId: string, manualText: string) => void;
-  hasInvalidManualResolution: boolean;
-  handleResolveConflicts: () => Promise<void>;
-} {
-  const [conflictForm, setConflictForm] = React.useState<
-    Record<
-      string,
-      {
-        resolution: "ours" | "theirs" | "manual";
-        manualText: string;
-      }
-    >
-  >({});
-
-  React.useEffect(() => {
-    if (mergeConflicts.length === 0) {
-      setConflictForm({});
-      return;
-    }
-    setConflictForm((prev) => {
-      const next: Record<
-        string,
-        {
-          resolution: "ours" | "theirs" | "manual";
-          manualText: string;
-        }
-      > = {};
-      for (const conflict of mergeConflicts) {
-        next[conflict.conflictId] = prev[conflict.conflictId] ?? {
-          resolution: "ours",
-          manualText: "",
-        };
-      }
-      return next;
-    });
-  }, [mergeConflicts]);
-
-  const handleResolutionChange = React.useCallback(
-    (conflictId: string, resolution: "ours" | "theirs" | "manual") => {
-      setConflictForm((prev) => ({
-        ...prev,
-        [conflictId]: {
-          resolution,
-          manualText: prev[conflictId]?.manualText ?? "",
-        },
-      }));
-    },
-    [],
-  );
-
-  const handleManualTextChange = React.useCallback(
-    (conflictId: string, manualText: string) => {
-      setConflictForm((prev) => ({
-        ...prev,
-        [conflictId]: {
-          resolution: prev[conflictId]?.resolution ?? "manual",
-          manualText,
-        },
-      }));
-    },
-    [],
-  );
-
-  const hasInvalidManualResolution = React.useMemo(
-    () =>
-      mergeConflicts.some((conflict) => {
-        const selected = conflictForm[conflict.conflictId];
-        if (selected?.resolution !== "manual") {
-          return false;
-        }
-        return selected.manualText.trim().length === 0;
-      }),
-    [conflictForm, mergeConflicts],
-  );
-
-  const handleResolveConflicts = React.useCallback(async () => {
-    if (!documentId || !mergeSessionId) {
-      return;
-    }
-    if (hasInvalidManualResolution) {
-      return;
-    }
-
-    const resolutions = mergeConflicts.map((conflict) => {
-      const selected = conflictForm[conflict.conflictId] ?? {
-        resolution: "ours" as const,
-        manualText: "",
-      };
-      if (selected.resolution === "manual") {
-        return {
-          conflictId: conflict.conflictId,
-          resolution: "manual" as const,
-          manualText: selected.manualText.trim(),
-        };
-      }
-      return {
-        conflictId: conflict.conflictId,
-        resolution: selected.resolution,
-      };
-    });
-
-    const result = await resolveBranchConflict({
-      documentId,
-      mergeSessionId,
-      resolutions,
-      resolvedBy: "user",
-    });
-
-    if (result.ok) {
-      setConflictForm({});
-    }
-  }, [
-    conflictForm,
-    documentId,
-    hasInvalidManualResolution,
-    mergeConflicts,
-    mergeSessionId,
-    resolveBranchConflict,
-  ]);
-
-  return {
-    conflictForm,
-    handleResolutionChange,
-    handleManualTextChange,
-    hasInvalidManualResolution,
-    handleResolveConflicts,
-  };
-}
 type VersionHistoryContainerProps = {
   projectId: string;
 };
@@ -653,109 +243,22 @@ export function VersionHistoryContainer(
         showAiMarks={showAiMarks}
         showCloseButton={false}
       />
-      <div className="mx-3 mt-3 space-y-2 rounded-lg border border-[var(--color-border-default)] bg-[var(--color-bg-raised)] p-3">
-        <div className="text-xs font-semibold text-[var(--color-fg-default)]">
-          {t("versionHistory.container.branchMerge")}
-        </div>
-        <div className="grid grid-cols-1 gap-2">
-          {/* eslint-disable-next-line creonow/no-native-html-element -- specialized label */}
-          <label className="flex flex-col gap-1 text-[11px] text-[var(--color-fg-muted)]">
-            {t("versionHistory.container.sourceBranch")}
-            {/* eslint-disable-next-line creonow/no-native-html-element -- specialized input */}
-            <input
-              data-testid="branch-merge-source-input"
-              className="rounded border border-[var(--color-border-default)] bg-[var(--color-bg-base)] px-2 py-1 text-xs text-[var(--color-fg-default)]"
-              value={sourceBranchName}
-              onChange={(event) => setSourceBranchName(event.target.value)}
-            />
-          </label>
-          {/* eslint-disable-next-line creonow/no-native-html-element -- specialized label */}
-          <label className="flex flex-col gap-1 text-[11px] text-[var(--color-fg-muted)]">
-            {t("versionHistory.container.targetBranch")}
-            {/* eslint-disable-next-line creonow/no-native-html-element -- specialized input */}
-            <input
-              data-testid="branch-merge-target-input"
-              className="rounded border border-[var(--color-border-default)] bg-[var(--color-bg-base)] px-2 py-1 text-xs text-[var(--color-fg-default)]"
-              value={targetBranchName}
-              onChange={(event) => setTargetBranchName(event.target.value)}
-            />
-          </label>
-        </div>
-        {/* eslint-disable-next-line creonow/no-native-html-element -- specialized button */}
-        <button
-          type="button"
-          data-testid="branch-merge-submit"
-          className="focus-ring rounded border border-[var(--color-accent)] px-2 py-1 text-xs text-[var(--color-fg-default)] disabled:opacity-50"
-          onClick={() => void handleMergeBranches()}
-          disabled={
-            branchMergeStatus === "loading" ||
-            sourceBranchName.trim().length === 0 ||
-            targetBranchName.trim().length === 0
-          }
-        >
-          {branchMergeStatus === "loading"
-            ? t("versionHistory.container.merging")
-            : t("versionHistory.container.mergeBranches")}
-        </button>
-        {branchMergeStatus === "ready" ? (
-          <div className="text-[11px] text-[var(--color-success)]">
-            {t("versionHistory.container.branchMergeCompleted")}
-          </div>
-        ) : null}
-        {branchMergeStatus === "error" && branchMergeError ? (
-          <div role="alert" className="text-[11px] text-[var(--color-error)]">
-            {getHumanErrorMessage(branchMergeError)}
-          </div>
-        ) : null}
-      </div>
-      {branchMergeStatus === "conflict" && mergeConflicts.length > 0 ? (
-        <div
-          data-testid="branch-conflict-panel"
-          className="mx-3 mt-3 space-y-3 rounded-lg border border-[var(--color-warning)] bg-[var(--color-bg-raised)] p-3"
-        >
-          <div className="flex items-center justify-between gap-2">
-            <div className="text-xs font-semibold text-[var(--color-fg-default)]">
-              {t("versionHistory.container.mergeConflictDiff")}
-            </div>
-            {/* eslint-disable-next-line creonow/no-native-html-element -- specialized button */}
-            <button
-              type="button"
-              className="focus-ring rounded border border-[var(--color-border-default)] px-2 py-1 text-[11px] text-[var(--color-fg-muted)]"
-              onClick={clearBranchMergeState}
-            >
-              {t("versionHistory.container.dismiss")}
-            </button>
-          </div>
-
-          {mergeConflicts.map((conflict) => {
-            const selected = conflictForm[conflict.conflictId] ?? {
-              resolution: "ours" as const,
-              manualText: "",
-            };
-            return (
-              <BranchConflictItem
-                key={conflict.conflictId}
-                conflict={conflict}
-                selected={selected}
-                onResolutionChange={handleResolutionChange}
-                onManualTextChange={handleManualTextChange}
-                t={t}
-              />
-            );
-          })}
-
-          {/* eslint-disable-next-line creonow/no-native-html-element -- specialized button */}
-          <button
-            type="button"
-            data-testid="branch-conflict-submit"
-            className="focus-ring rounded border border-[var(--color-accent)] px-2 py-1 text-xs text-[var(--color-fg-default)] disabled:opacity-50"
-            onClick={() => void handleResolveConflicts()}
-            disabled={hasInvalidManualResolution}
-          >
-            {t("versionHistory.container.submitConflictResolution")}
-          </button>
-        </div>
-      ) : null}
+      <BranchMergeSection
+        sourceBranchName={sourceBranchName}
+        onSourceChange={setSourceBranchName}
+        targetBranchName={targetBranchName}
+        onTargetChange={setTargetBranchName}
+        branchMergeStatus={branchMergeStatus}
+        branchMergeError={branchMergeError}
+        onMergeBranches={() => void handleMergeBranches()}
+        mergeConflicts={mergeConflicts}
+        conflictForm={conflictForm}
+        onResolutionChange={handleResolutionChange}
+        onManualTextChange={handleManualTextChange}
+        hasInvalidManualResolution={hasInvalidManualResolution}
+        onResolveConflicts={() => void handleResolveConflicts()}
+        onDismissConflicts={clearBranchMergeState}
+      />
       {previewStatus === "error" && previewError ? (
         <div
           role="alert"
